AX: Implement the AXSearchManager class. https://bugs.webkit.org/show_bug.cgi?id=276684 <rdar://problem/131867309>
Reviewed by Tyler Wilcock.

This new class encapsulates the findMatchingObject functionality that was previously in the Accessibility namespace. In addition, it allows to keep state for ranges within objects which is necessary for navigating by misspellings which clients like VoiceOver use. This new functionality will be further used in a subsequent patch. Some code cleanup.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/accessibility/AXCoreObject.h:
(WebCore::AccessibilitySearchCriteria::AccessibilitySearchCriteria): Deleted.
* Source/WebCore/accessibility/AXLogger.cpp:
* Source/WebCore/accessibility/AXSearchManager.cpp: Added.
(WebCore::isRadioButtonInDifferentAdhocGroup):
(WebCore::AXSearchManager::matchForSearchKeyAtIndex):
(WebCore::AXSearchManager::match):
(WebCore::AXSearchManager::matchText):
(WebCore::AXSearchManager::matchWithResultsLimit):
(WebCore::appendAccessibilityObject):
(WebCore::appendChildrenToArray):
(WebCore::AXSearchManager::findMatchingObjects):
* Source/WebCore/accessibility/AXSearchManager.h: Added.
(WebCore::AccessibilitySearchCriteria::AccessibilitySearchCriteria):
* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::spellCheckerResults const):
(WebCore::AccessibilityObject::addChild):
(WebCore::AccessibilityObject::findMatchingObjects):
(WebCore::AccessibilityObject::hasMisspelling const): Deleted.
(WebCore::appendAccessibilityObject): Deleted.
(WebCore::Accessibility::isRadioButtonInDifferentAdhocGroup): Deleted.
(WebCore::Accessibility::isAccessibilityObjectSearchMatchAtIndex): Deleted.
(WebCore::Accessibility::isAccessibilityObjectSearchMatch): Deleted.
(WebCore::Accessibility::isAccessibilityTextSearchMatch): Deleted.
(WebCore::Accessibility::objectMatchesSearchCriteriaWithResultLimit): Deleted.
(WebCore::Accessibility::appendChildrenToArray): Deleted.
(WebCore::Accessibility::findMatchingObjects): Deleted.
* Source/WebCore/accessibility/AccessibilityObject.h:
* Source/WebCore/accessibility/ios/WebAccessibilityObjectWrapperIOS.mm:
(-[WebAccessibilityObjectWrapper accessibilityFindMatchingObjects:]):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::findMatchingObjects):
(WebCore::AXIsolatedObject::spellCheckerResults const):
(WebCore::AXIsolatedObject::hasMisspelling const): Deleted.
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.h:
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperBase.mm:
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper accessibilityAttributeValue:forParameter:]):

Canonical link: https://commits.webkit.org/281058@main
