+++ Source/WebCore/Headers.cmake
@@ -657,6 +657,7 @@ set(WebCore_PRIVATE_FRAMEWORK_HEADERS
     accessibility/AXGeometryManager.h
     accessibility/AXLogger.h
     accessibility/AXObjectCache.h
+    accessibility/AXSearchManager.h
     accessibility/AXTextMarker.h
     accessibility/AXTextRun.h
     accessibility/AXTextStateChangeIntent.h
